Fix CI: 32-bit seed overflow, bare-score column regression
Cross-compile 386 failed: FNV seed 2654435761 overflows int on 32-bit.
Retype the test generators through uint32/uint64 arithmetic that is
bit-identical on 64-bit and wraps deterministically on 32-bit.

Gateway TestCoordinatorRejectsDeclaredDomainMismatchBeforeSourceIO (and
race distributed) failed: kwScore marking excluded a bare score column
from the nameable-word lane in comparison operands and alias rejection,
so WHERE id=score died at parse instead of reaching schema analysis.
Passes on main; restores parity.

+++ internal/tin/score_topk_test.go
@@ -17,7 +17,7 @@ func topKExactIndex(t *testing.T, seal bool, n int) *Index {
 		var sb strings.Builder
 		sb.WriteString("common filler")
 		if i%5 != 0 {
-			tf := 1 + (i*2654435761)%512
+			tf := 1 + int(uint64(i)*2654435761%512)
 			for k := 0; k < tf; k++ {
 				sb.WriteString(" zipf")
 			}
@@ -104,7 +104,7 @@ func skewTopKIndex(t testing.TB, seal bool) *Index {
 		var sb strings.Builder
 		sb.WriteString("fill")
 		if i%5 != 0 {
-			tf := 1 + (i*2654435761)%1024
+			tf := 1 + int(uint64(i)*2654435761%1024)
 			for k := 0; k < tf; k++ {
 				sb.WriteString(" zipf")
 			}
